backend/requirements.txt
flask
flask-cors
python-dotenv
googlemaps
requests
gunicorn
networkx
openai
# psycopg2